Revert "forgot to change it to long in actual cassandra table"
This reverts commit e12f44034159556862c5748032142d64da9287f2.

+++ heat-gen-processor/src/com/nestedtori/heatgen/TempConsumer.java
@@ -88,7 +88,7 @@ public class TempConsumer implements Runnable {
 		        	TimeTempTuple value = record.value();
 		        	double x = leftX + k.i * dx;
 		     		double y = bottomY + k.j * dy;
-		     		BoundStatement bs = ps.bind((int)(value.time/timeUnit), x, y, value.val);
+		     		BoundStatement bs = ps.bind(value.time/timeUnit, x, y, value.val);
 		     		batch.add(bs);	     		
 		         }
 		         session.executeAsync(batch);
